backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
pymongo==4.6.1
elasticsearch==8.11.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyjwt==2.8.0
cryptography==41.0.7
google-generativeai==0.3.1
requests==2.31.0
ldap3==2.9.1
apscheduler==3.10.4
python-dotenv==1.0.0
python-json-logger==2.0.7
gunicorn==21.2.0
python-dateutil==2.8.2
pytest==7.4.3
pytest-cov==4.1.0
